requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
boto3==1.34.0
django-storages==1.14.2
twilio==8.10.0
pyotp==2.9.0
django-filter==23.5
drf-yasg==1.21.7
django-extensions==3.2.3
python-dotenv==1.0.0
pg8000==1.30.4
requests==2.31.0
python-binance==1.0.19
razorpay==1.4.1

# Blockchain Libraries for Real Wallets
web3==6.11.0
tronpy==0.4.0
eth-account==0.9.0
cryptography>=41.0.0
mnemonic>=0.20

# Testing Dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.2.2
pytest-mock==3.12.0
pytest-xdist==3.3.1